requirements.txt
prompt-toolkit
rainbow_tqdm
requests
aiohttp
PyYAML
types-aiofiles
aiofiles
tenacity